+++ changelog.md
@@ -1,4 +1,4 @@
-# Version 0.1a
+# Version 0.1.1
 
 ## New Features
 
@@ -9,3 +9,5 @@
 
 - Preferences are saved as QSettings
 - Set open/save-path to homefolder
+- Fixed: Message after filewrite to filesystem (e.g. odt-Export)
+- Fixed: Change to multiple Arguments